+++ Cesium3DTilesSelection/src/CreditSystem.cpp
@@ -4,7 +4,7 @@
 
 namespace Cesium3DTilesSelection {
 
-Credit CreditSystem::createCredit(const std::string& html) {
+Credit CreditSystem::createCredit(const std::string& html, bool showOnScreen) {
   // if this credit already exists, return a Credit handle to it
   for (size_t id = 0; id < _credits.size(); ++id) {
     if (_credits[id].html == html) {
@@ -13,10 +13,17 @@ Credit CreditSystem::createCredit(const std::string& html) {
   }
 
   // this is a new credit so add it to _credits
-  _credits.push_back({html, -1});
+  _credits.push_back({html, showOnScreen, -1});
 
   // return a Credit handle to the newly created entry
   return Credit(_credits.size() - 1);
+}
+
+const bool CreditSystem::shouldBeShownOnScreen(Credit credit) const noexcept {
+  if (credit.id < _credits.size()) {
+    return _credits[credit.id].showOnScreen;
+  }
+  return false;
 }
 
 const std::string& CreditSystem::getHtml(Credit credit) const noexcept {